Cleanup item: Create BVC+TabToolbarDelegate.swift
